+++ src/lib/components/helpers/Tooltip.tsx
@@ -16,13 +16,15 @@ const Tooltip = ({ children, title, placement = 'top' }: ITooltipProps) => {
     >
       {title && (
         <div
-          className={`absolute flex justify-start items-center py-1.5 px-2.5 rounded-md border border-black dark:border-white z-50 bg-white dark:bg-black pointer-events-none select-none',
-            ${visible ? 'opacity-100' : 'opacity-0'}
-            ${placement === 'right' ? 'left-14' : ''}
-            ${placement === 'left' ? 'right-14' : ''}
-            ${placement === 'top' ? 'bottom-10' : ''}`}
+          className={`efw-absolute efw-flex efw-justify-start efw-items-center efw-py-1.5 efw-px-2.5 efw-rounded-md efw-border efw-border-black dark:efw-border-white efw-z-50 efw-bg-white dark:efw-bg-black efw-pointer-events-none efw-select-none',
+            ${visible ? 'efw-opacity-100' : 'efw-opacity-0'}
+            ${placement === 'right' ? 'efw-left-14' : ''}
+            ${placement === 'left' ? 'efw-right-14' : ''}
+            ${placement === 'top' ? 'efw-bottom-10' : ''}`}
         >
-          <span className={'text-[12px] w-max text-start max-w-40'}>
+          <span
+            className={'efw-text-[12px] efw-w-max efw-text-start efw-max-w-40'}
+          >
             {title}
           </span>
         </div>