+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 0.5.0 : TBD
+## 0.5.1 : 2024-07-16
+
+- **Added**: prefix to tailwind
+- **Added**: Responsivness to edit
+- **Added**: Resize when getting layer to adjust given elements
+- **Changed**: edit `save layer` functions
+- **Removed**: elements useState
+- **Fixed**: stroke width responsivness
+
+## 0.5.0 : 2024-07-16
 
 - **Added**: Edit mode
 - **Added**: New params to viewer